requirements/base.txt
Django==3.0.8
Pillow==7.2.0
psycopg2-binary==2.8.5
pytz==2020.1
django-environ==0.4.5
Markdown==3.2.2
django-rq==2.3.2
pyaml==20.4.0
